feat: Add rake task to fix undefined areas for manually added site addresses
This commit adds a new rake task `undefined_area_fix` to fix areas for manually added site addresses. The task finds registrations with undefined EA (Environment Agency) area and attempts to fix them by matching the postcode with the corresponding area. If a match is found, the area is updated for the site address. The task also provides statistics on the number of registrations with undefined EA area and the number of addresses that were fixed or couldn't be fixed.