[RUBY-3297] feat: Add download functionality to BulkExportsController
This commit adds a new action `download` to the `BulkExportsController` that allows users to download generated reports. The action authorizes the user, retrieves the generated report, generates a presigned URL for the report file, tracks the download, and redirects the user to the download URL. This enhances the functionality of the `BulkExportsController` by providing a convenient way for users to download their reports.