RUBY-3297: Ability to track report downloads + export download history (#1473)
* [RUBY-3297] Add Reports::Download model and migration

* [RUBY-3297] feat: Add Reports::Download service and spec

This commit adds a new service called `Reports::TrackDownloadService` along with its corresponding spec file. The service is responsible for creating a download record with the correct attributes, such as the report, user ID, and downloaded timestamp. The spec file includes tests to ensure the service functions as expected, including handling cases where the user is nil.

#

* [RUBY-3297] feat: Add download functionality to BulkExportsController

This commit adds a new action `download` to the `BulkExportsController` that allows users to download generated reports. The action authorizes the user, retrieves the generated report, generates a presigned URL for the report file, tracks the download, and redirects the user to the download URL. This enhances the functionality of the `BulkExportsController` by providing a convenient way for users to download their reports.

* [RUBY-3297] db schema update

* [RUBY-3297] refactoring Reports::Download model to store report file_name instead of report id

* [RUBY-3297] refactoring TrackDownloadService to store report file_name instead of report id

* [RUBY-3297] adding DownloadHistoryExportService

* [RUBY-3297] Ability to export download history into CSV

* [RUBY-3297] adding specs for ability to view report download history by system user