[RUBY-3720] Refactor RefreshCompaniesHouseNameService to set reason for change
This commit refactors the RefreshCompaniesHouseNameService to set the reason for change when refreshing the Companies House information. The `registration.reason_for_change` attribute is updated with the value from the locale file, `refresh_companies_house_name.en.yml`. This change ensures that the reason for change is properly recorded when the Companies House information is refreshed.